backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.4.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
tensorflow>=2.15.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv==1.0.1
